tests/krate/downloads: Move tests into `routes` submodules
