+++ src/tests/routes/crates/mod.rs
@@ -1,1 +1,2 @@
+pub mod downloads;
 pub mod versions;
