+++ src/tests/routes/crates/versions/mod.rs
@@ -1,1 +1,2 @@
 pub mod dependencies;
+pub mod download;
